[HIGH][REFACTOR] Split StreamControlsWidget.tsx into smaller components
Refactored the 2,225-line StreamControlsWidget.tsx into 6 focused components:
- PreStreamSettings.tsx (~235 lines) - Pre-stream configuration UI
- ActivityRow.tsx (~240 lines) - Individual activity feed items
- StreamActionsPanel.tsx (~95 lines) - Stream action buttons
- TimersPanel.tsx (~235 lines) - Timer management UI
- LiveStreamControlCenter.tsx (~770 lines) - Live stream control center
- PostStreamSummary.tsx (~133 lines) - Post-stream stats summary

Also created shared types.ts file for all types, constants, schemas, and helper functions.

The main StreamControlsWidget.tsx is now ~130 lines and uses composition pattern.
All components are re-exported for backwards compatibility.
